Renamed the logging code files to all lowercase names
