+++ include/pistache/meson.build
@@ -29,10 +29,10 @@ install_headers(
 	'net.h',
 	'os.h',
 	'peer.h',
-        'PistCheck.h',
-        'PistSysLog.h',
+        'pist_check.h',
+        'pist_syslog.h',
 	'prototype.h',
-        'PS_TimeLog.h',
+        'pist_timelog.h',
 	'reactor.h',
 	'route_bind.h',
 	'router.h',
