+++ include/pistache/pist_check.h
@@ -5,7 +5,7 @@
  */
 
 /******************************************************************************
- * PistCheck.h
+ * pist_check.h
  * 
  * Debugging breakpoints
  *
@@ -14,7 +14,7 @@
 #ifndef INCLUDED_PSCHECK_H
 #define INCLUDED_PSCHECK_H
 
-#include "PistSysLog.h"
+#include "pist_syslog.h"
 
 // If DEBUG is enabled, PS_LOGDBG_STACK_TRACE logs a stack trace
 #ifdef DEBUG
