+++ src/meson.build
@@ -15,9 +15,9 @@ pistache_common_src = [
 	'common'/'net.cc',
 	'common'/'os.cc',
 	'common'/'peer.cc',
-        'common'/'PistCheck.cc',
-        'common'/'PistSysLog.cc',
-        'common'/'PS_TimeLog.cc',
+        'common'/'pist_check.cc',
+        'common'/'pist_syslog.cc',
+        'common'/'pist_timelog.cc',
 	'common'/'reactor.cc',
 	'common'/'stream.cc',
 	'common'/'string_logger.cc',
